수정 : 장바구니 전체삭제 - 유저id 사용하여 장바구니dao 로 존재 여부 확인

+++ services/cartService.js
@@ -60,7 +60,15 @@ const deleteCart = async (cartId) => {
 }
 
 const deleteAllCart = async (userId) => {
-    await cartDao.deleteAllCartByUserId(userId);
+
+    const existCart = await cartDao.getCartByUserId(userId)
+
+    if (existCart.length === 0) {
+        throw {status : 404, message : "CART_DOES_NOT_EXIST"}
+    } else {
+        await cartDao.deleteAllCartByUserId(userId);
+        return "CART_DELETED_SUCCESSFULLY"
+    }
 }
 
 module.exports = {
